After chat with Mathieu Lauriere
Some ideas:
-New points at every iteration
-Lower the value of the potential
-Transfer Learning with increasing the potential value

The first thing would be to try change the architecture in something simpler (and deeper). Like 5 layers of FNN with 200 neurons each. One important thing is to have the last layer to be a ReLU layer, so we don't have negative solutions. The activation function in the hidden layers could be tanh or sigmoid. (I'm using sigmoid)

Then we need to change the way we sample the points: we start with constant warmstart with very very few points (like 1000) and then start feeding 100 more every iteration. 
We are also increasing the points considered in the adaptive resampling (the pool from which to choose the worst performing 100 points is bigger).